.history/requirements_20250303110350.txt
numpy
opencv-python
trimesh
pyrender
Pillow
torch
ultralytics 